refactor: обновлены пути к файлам feat: добавлена работа с порогами изменения сенсорных каналов
